+++ service-product/module-batch/src/main/java/com/github/msafriends/modulebatch/config/strategy/QueryStrategy.java
@@ -1,0 +1,5 @@
+package com.github.msafriends.modulebatch.config.strategy;
+
+public interface QueryStrategy {
+    String getInsertSQL();
+}
