--- service-product/module-batch/src/main/java/src/main/java/com/github/msafriends/modulebatch/config/strategy/QueryStrategy.java
@@ -1,5 +1,0 @@
-package src.main.java.com.github.msafriends.modulebatch.config.strategy;
-
-public interface QueryStrategy {
-    String getInsertSQL();
-}
